SiloPlayer: gate over-video glass off low-power TVs, migrate to siloGlass
Opening the player menu spiked CPU on the A12 Apple TV: the HUD panels
(.regularMaterial) and control capsules (.ultraThinMaterial) are
backdrop-sampling effects, so the render server re-samples and
re-blurs the covered video region on every video frame for as long as
the menu is up — pure per-frame overhead on a 4K surface.

All player-over-video glass (HUD panels, control capsules, hold-seek
pill, notice overlay, PlayerView chrome) now routes through one
siloPlayerGlass modifier: low-power Apple TVs draw a non-sampling
translucent fill (visually close over dark video, free), everything
else renders Silo's Liquid Glass via the existing siloGlass path —
which also moves the player off the last legacy materials.

The AppleTV14,x cutoff moves into a shared DevicePower helper, now
used by both this gate and the subtitle raster cap.

490 tests pass; iOS/tvOS/macOS build.
