+++ iosApp/iosApp/DesignSystem/SiloGlass.swift
@@ -11,4 +11,20 @@ extension View {
         if interactive { glass = glass.interactive() }
         return self.glassEffect(glass, in: shape)
     }
+
+    /// Glass for surfaces drawn over LIVE VIDEO (player HUD, controls,
+    /// notices). Backdrop-sampling effects — glassEffect and the legacy
+    /// materials alike — make the render server re-sample and re-blur the
+    /// covered video region on every video frame, which A12-class Apple
+    /// TVs pay for as a visible spike whenever the player menu is up.
+    /// Low-power devices draw a non-sampling translucent fill instead;
+    /// everything else gets standard Silo glass.
+    @ViewBuilder
+    func siloPlayerGlass(in shape: some Shape, tint: Color? = nil) -> some View {
+        if DevicePower.isLowPowerAppleTV {
+            self.background(shape.fill(Color(white: 0.10).opacity(0.88)))
+        } else {
+            self.siloGlass(in: shape, tint: tint)
+        }
+    }
 }
