+++ iosApp/iosApp/Screens/Player/tvOS/TVPlayerControls.swift
@@ -176,7 +176,7 @@ struct TVPlayerControls: View {
                 }
                 .padding(.horizontal, 12)
                 .padding(.vertical, 6)
-                .background(Capsule().fill(.ultraThinMaterial))
+                .siloPlayerGlass(in: Capsule())
             }
 
             if viewModel.sleepTimer.isActive {
@@ -187,7 +187,7 @@ struct TVPlayerControls: View {
                     .monospacedDigit()
                     .padding(.horizontal, 12)
                     .padding(.vertical, 6)
-                    .background(Capsule().fill(.ultraThinMaterial))
+                    .siloPlayerGlass(in: Capsule())
             }
         }
     }
